requirements.txt
backports.zoneinfo==0.2.1
boto3==1.22.10
botocore==1.25.10
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
Flask==2.1.2
Flask-Session==0.4.0
idna==3.3
image==1.5.33
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
Pillow==9.1.0
PyJWT==2.3.0
pymongo==4.1.1
python-dateutil==2.8.2
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
